Feature(APPC-4467) Remove unused code and refactor home to correct packaged
