--- app/src/main/java/com/asfoundation/wallet/home/ui/list/transactions/empty/EmptyItem.kt
@@ -1,8 +1,0 @@
-package com.asfoundation.wallet.home.ui.list.transactions.empty
-
-data class EmptyItem(
-  val id: String,
-  val animationRes: Int,
-  val titleText: String,
-  val bodyText: String
-)
