--- app/src/main/java/com/asfoundation/wallet/home/usecases/SetSeenFingerprintTooltipUseCase.kt
@@ -1,13 +1,0 @@
-package com.asfoundation.wallet.home.usecases
-
-import com.appcoins.wallet.sharedpreferences.FingerprintPreferencesDataSource
-import javax.inject.Inject
-
-class SetSeenFingerprintTooltipUseCase @Inject constructor(
-  private val fingerprintPreferences: FingerprintPreferencesDataSource
-) {
-
-  operator fun invoke() {
-    fingerprintPreferences.setSeenFingerprintTooltip()
-  }
-}
