--- app/src/main/java/com/asfoundation/wallet/home/usecases/UpdateTransactionsNumberUseCase.kt
@@ -1,30 +1,0 @@
-package com.asfoundation.wallet.home.usecases
-
-import com.asfoundation.wallet.rating.RatingRepository
-import com.asfoundation.wallet.transactions.Transaction
-import javax.inject.Inject
-
-class UpdateTransactionsNumberUseCase @Inject constructor(
-  private val ratingRepository: RatingRepository
-) {
-
-  companion object {
-    const val MINIMUM_TRANSACTIONS_NR = 5
-  }
-
-  operator fun invoke(transactions: List<Transaction>) {
-    var transactionsNumber = 0
-    for (transaction in transactions) {
-      if ((transaction.type == Transaction.TransactionType.IAP
-            || transaction.type == Transaction.TransactionType.TOP_UP
-            || transaction.type == Transaction.TransactionType.IAP_OFFCHAIN)
-        && transaction.status == Transaction.TransactionStatus.SUCCESS
-      ) {
-        if (++transactionsNumber >= MINIMUM_TRANSACTIONS_NR && !ratingRepository.hasEnoughSuccessfulTransactions()) {
-          ratingRepository.saveEnoughSuccessfulTransactions()
-          break
-        }
-      }
-    }
-  }
-}
