--- app/src/main/java/com/asfoundation/wallet/promotions/ui/list/PromotionClick.kt
@@ -1,6 +1,0 @@
-package com.asfoundation.wallet.promotions.ui.list
-
-data class PromotionClick(
-  val id: String,
-  val extras: Map<String, String>? = null
-)
