--- app/src/main/java/com/asfoundation/wallet/promotions/ui/list/model/EmptyView.kt
@@ -1,17 +1,0 @@
-package com.asfoundation.wallet.promotions.ui.list.model
-
-import android.content.Context
-import android.util.AttributeSet
-import android.widget.FrameLayout
-import com.airbnb.epoxy.ModelView
-
-@ModelView(autoLayout = ModelView.Size.WRAP_WIDTH_WRAP_HEIGHT)
-class EmptyView : FrameLayout {
-
-  constructor(context: Context) : this(context, null)
-  constructor(context: Context, attrs: AttributeSet?) : this(context, attrs, 0)
-  constructor(context: Context, attrs: AttributeSet?, defStyleAttr: Int) : super(
-    context, attrs,
-    defStyleAttr
-  )
-}
