--- app/src/main/res/layout/fragment_my_wallets.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/styleguide_blue"
-    android:orientation="vertical"
-    >
-  <include
-      android:id="@+id/toolbar"
-      layout="@layout/fragment_my_wallets_toolbar"
-      />
-  <ScrollView
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      android:clipToPadding="false"
-      android:paddingTop="16dp"
-      >
-    <include
-        android:id="@+id/my_wallets_content"
-        layout="@layout/fragment_my_wallets_content"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        />
-  </ScrollView>
-</LinearLayout>
