--- app/src/main/res/layout/fragment_my_wallets_name.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    android:paddingHorizontal="16dp"
-    android:paddingVertical="24dp"
-    >
-  <TextView
-      android:id="@+id/title"
-      style="@style/MyWallets.Title.Dark"
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:text="@string/mywallet_choose_name_header"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      />
-  <ProgressBar
-      android:id="@+id/loading"
-      android:layout_width="44dp"
-      android:layout_height="44dp"
-      android:layout_marginVertical="48dp"
-      android:indeterminateDrawable="@drawable/gradient_progress"
-      android:visibility="visible"
-      app:layout_constraintBottom_toBottomOf="parent"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toTopOf="parent"
-      />
-  <com.appcoins.wallet.ui.widgets.WalletTextFieldView
-      android:id="@+id/name_input"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:layout_marginTop="16dp"
-      android:maxHeight="170dp"
-      app:layout_constraintStart_toStartOf="parent"
-      app:layout_constraintTop_toBottomOf="@id/title"
-      app:textFieldColor="@color/styleguide_white"
-      app:textFieldHint="@string/mywallet_choose_name_field"
-      app:textFieldType="outlined"
-      />
-  <com.appcoins.wallet.ui.widgets.WalletButtonView
-      android:id="@+id/save"
-      android:layout_width="wrap_content"
-      android:layout_height="@dimen/large_button_height"
-      android:layout_marginTop="24dp"
-      android:layout_marginEnd="@dimen/big_margin"
-      app:buttonText="@string/action_save"
-      app:buttonType="filled"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintTop_toBottomOf="@+id/name_input"
-      />
-  <com.appcoins.wallet.ui.widgets.WalletButtonView
-      android:id="@+id/cancel"
-      android:layout_width="wrap_content"
-      android:layout_height="@dimen/large_button_height"
-      android:layout_marginTop="24dp"
-      android:layout_marginEnd="@dimen/big_margin"
-      app:buttonColor="@color/styleguide_dark_grey"
-      app:buttonText="@string/cancel_button"
-      app:buttonType="text"
-      app:layout_constraintEnd_toStartOf="@+id/save"
-      app:layout_constraintTop_toBottomOf="@+id/name_input"
-      />
-</androidx.constraintlayout.widget.ConstraintLayout>
