--- app/src/main/res/layout/fragment_my_wallets_toolbar.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="56dp"
-    android:clickable="true"
-    android:focusable="true"
-    android:gravity="center_vertical"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toTopOf="parent"
-    tools:background="@color/styleguide_blue"
-    >
-  <ImageView
-      android:id="@+id/toolbar_layout_logo"
-      android:layout_width="67dp"
-      android:layout_height="24dp"
-      android:layout_margin="16dp"
-      android:contentDescription="@null"
-      android:src="@drawable/ic_app_logo"
-      />
-  <View
-      android:layout_width="0dp"
-      android:layout_height="match_parent"
-      android:layout_weight="1"
-      />
-  <ImageButton
-      android:id="@+id/action_button_more"
-      style="@style/Widget.AppCompat.ActionButton"
-      android:layout_width="42dp"
-      android:layout_height="48dp"
-      android:layout_marginEnd="7dp"
-      android:contentDescription="@null"
-      android:src="@drawable/ic_more"
-      app:tint="@color/styleguide_white"
-      tools:ignore="SpeakableTextPresentCheck"
-      />
-</LinearLayout>
