--- app/src/main/res/layout/fragment_promotions.xml
@@ -1,154 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
-
-  <com.google.android.material.appbar.AppBarLayout
-      android:id="@+id/promotions_header_appbar"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      >
-
-    <com.google.android.material.appbar.CollapsingToolbarLayout
-        android:id="@+id/collapsing_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/styleguide_blue"
-        android:minHeight="?attr/actionBarSize"
-        app:layout_scrollFlags="scroll|exitUntilCollapsed"
-        app:titleEnabled="false"
-        >
-
-      <include
-          android:id="@+id/current_level_header"
-          layout="@layout/promotions_gamification_header"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:layout_marginTop="?attr/actionBarSize"
-          android:animateLayoutChanges="true"
-          app:layout_collapseMode="parallax"
-          app:layout_collapseParallaxMultiplier="1"
-          />
-
-      <androidx.coordinatorlayout.widget.CoordinatorLayout
-          android:id="@+id/header_loading_container"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:layout_marginVertical="?attr/actionBarSize"
-          android:background="@color/styleguide_blue"
-          android:paddingVertical="@dimen/large_margin"
-          android:visibility="gone"
-          >
-        <com.airbnb.lottie.LottieAnimationView
-            android:id="@+id/header_loading"
-            android:layout_width="64dp"
-            android:layout_height="64dp"
-            android:layout_gravity="center_horizontal"
-            app:lottie_autoPlay="true"
-            app:lottie_enableMergePathsForKitKatAndAbove="true"
-            app:lottie_loop="true"
-            app:lottie_rawRes="@raw/loading_dark_animation"
-            />
-      </androidx.coordinatorlayout.widget.CoordinatorLayout>
-
-      <androidx.coordinatorlayout.widget.CoordinatorLayout
-          android:id="@+id/promotions_list_title_container"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:layout_gravity="bottom"
-          android:background="@drawable/background_active_wallet"
-          android:minHeight="28dp"
-          android:paddingHorizontal="@dimen/big_margin"
-          android:paddingTop="@dimen/big_margin"
-          app:layout_collapseMode="none"
-          >
-
-        <include
-            android:id="@+id/promotions_list_title_layout"
-            layout="@layout/item_promotions_title"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            />
-      </androidx.coordinatorlayout.widget.CoordinatorLayout>
-
-      <include
-          android:id="@+id/toolbar"
-          layout="@layout/fragment_promotions_toolbar"
-          android:layout_width="match_parent"
-          android:layout_height="?attr/actionBarSize"
-          app:layout_collapseMode="pin"
-          />
-    </com.google.android.material.appbar.CollapsingToolbarLayout>
-  </com.google.android.material.appbar.AppBarLayout>
-
-  <androidx.constraintlayout.widget.ConstraintLayout
-      android:id="@+id/promotions_content"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      android:background="@color/styleguide_white"
-      app:layout_behavior="@string/appbar_scrolling_view_behavior"
-      >
-
-    <include
-        android:id="@+id/promotions_skeleton"
-        layout="@layout/promotions_list_skeleton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        />
-
-    <com.airbnb.epoxy.EpoxyRecyclerView
-        android:id="@+id/rv_promotions"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:overScrollMode="never"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:listitem="@layout/item_promotions_default"
-        />
-
-    <include
-        android:id="@+id/no_network"
-        layout="@layout/no_network_retry_only_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        />
-
-    <include
-        android:id="@+id/no_promotions"
-        layout="@layout/no_promotions_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="@dimen/large_padding"
-        android:visibility="gone"
-        app:layout_constraintTop_toTopOf="parent"
-        />
-
-    <include
-        android:id="@+id/locked_promotions"
-        layout="@layout/locked_promotions_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        />
-  </androidx.constraintlayout.widget.ConstraintLayout>
-
-  <androidx.coordinatorlayout.widget.CoordinatorLayout
-      android:id="@+id/bottomsheet_coordinator_container"
-      android:layout_width="0dp"
-      android:layout_height="0dp"
-      android:background="@color/styleguide_black_transparent_80"
-      android:clickable="true"
-      android:elevation="2dp"
-      android:focusable="true"
-      android:visibility="gone"
-      />
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
