+++ app/src/main/res/navigation/home_graph.xml
@@ -6,6 +6,6 @@
 
   <fragment
       android:id="@+id/home_fragment"
-      android:name="com.asfoundation.wallet.wallet.home.HomeFragment" />
+      android:name="com.asfoundation.wallet.home.HomeFragment" />
 
 </navigation>
